requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# AI
google-generativeai==0.8.3
numpy==2.1.3
pandas==2.2.3

# HTTP / Scraping
python-jobspy==1.1.72
httpx==0.28.0
aiohttp==3.11.10
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
bleach==6.2.0
tenacity==9.0.0

# Resume Processing
pypdf2==3.0.1
python-docx==1.1.2

# Telegram (optional — used only for notifications)
python-telegram-bot==21.8

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.24.0
